backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
langchain==0.2.16
langgraph==0.2.19
langchain-google-genai==1.0.8
tavily-python==0.5.0
